Fix bug deleting received file on Windows
This was causing the following error with verify_file():

    WindowsError: [Error 32] The process cannot access the file because it is
    being used by another process:
    '[test-code-location]\GettingStartedTest.test_simple.received.txt'

The cause was that the received file was left open in verify_file when
its contents were passed to verify_files, which will remove the received
file if approved and received file are the same.

This was noticeable in Windows but not Linux.