fix crash when namer was configured to use a directory that didn't exist
When using a config file for the namer that specifies a directory to store
approved files in, it's useful if that directory is created if it doesn't
exist. I also converted the existing test for StringWriter to use pytest
and take advantage of the tmpdir fixture for cleaning up test files.