e   run tests with github actions
This commit will result in github acitons running our test suite on push and
pull requests. Tests are run across the same Python versions as the current
Travis CI config and are run on both ubuntu and macos. Tests on windows
failed but fixing up the Windows tests is for another day.

The .NET, C++ and Java libraries have started to use github actions for
various purposes. I think that it is worth trialing github actions
because:

* deployments seem more configurable, which should make the path to automatic
  deployments of this package smoother (I can't see a way of deploying to
  test.pypi.org from Travis CI).
* this test workflow has been configured to run on all pushes, which
  makes it easier for devs to test enhancements to the CI/CD process on
  their forks.

"GitHub Actions usage is free for public repositories and self-hosted
runners." see
https://docs.github.com/en/free-pro-team@latest/github/setting-up-and-managing-billing-and-payments-on-github/about-billing-for-github-actions#about-billing-for-github-actions