t   Fix test that fails on CI, all 3 platforms
This test was added 5 years ago, and passed because vimdiff
was included in the list of known reporters, and the CI was
only run on linux, which has vimdiff.

Later, PythonNative was added to GenericDiffReporter,
so when vimdiff was removed, the test passed by virtue of finding
a PythonNative "reporter".

Co-Authored-By: Llewellyn Falco <10874+isidore@users.noreply.github.com>