naming things and make code easier to understand
Make it clear what the function of the threading.local() object is (it is not the default reporter it just serves as a storage object for the reporter). The actual default reporter can also already be initialized at import time and reused, instead of creating a new instance every time (although in practice this not makes a difference, it is clearer that only one default reporter instance is needed).

Removal of global: the global keyword would only need to be used, if the module level name _THREAD_LOCAL_STORE should be reassigned. A only an attribute on the object is set though (reporter), the global statement is a no-op.

try except instead of hasattr check, etc.: I find this easier to read and using try/except like this as part of normal control flow is quite usual in Python.